chore: refactor graph introspect
