--- crates/rover-client/src/introspection/mod.rs
@@ -1,2 +1,0 @@
-mod schema;
-pub use schema::Schema;
