+++ crates/rover-client/src/query/graph/introspect/fixtures/interfaces.json
