+++ crates/rover-client/src/query/graph/introspect/fixtures/simple.json
