+++ crates/rover-client/src/query/graph/introspect/fixtures/swapi.json
