+++ crates/rover-client/src/query/graph/introspect/introspect_query.graphql
@@ -1,4 +1,4 @@
-query IntrospectionQuery {
+query GraphIntrospectQuery {
   __schema {
     queryType {
       name
