+++ crates/rover-client/src/query/graph/introspect/introspect_schema.graphql
