+++ crates/rover-client/src/query/graph/introspect/mod.rs
@@ -1,0 +1,7 @@
+mod query_runner;
+mod schema;
+mod types;
+
+pub use query_runner::run;
+pub use schema::Schema;
+pub use types::{GraphIntrospectInput, GraphIntrospectResponse};
